refactor: View models in Fan.Web #332
