+++ src/Core/Fan.Web/Models/AnalyticsVM.cs
@@ -1,6 +1,6 @@
-﻿namespace Fan.Web.Models.Home
+﻿namespace Fan.Web.Models
 {
-    public class AnalyticsViewModel
+    public class AnalyticsVM
     {
         public string GoogleAnalyticsTrackingID { get; set; }
         public string AppInsightsFullScript { get; set; }
