+++ src/Core/Fan.Web/Models/CookieConsentVM.cs
@@ -1,6 +1,6 @@
-﻿namespace Fan.Web.Models.Home
+﻿namespace Fan.Web.Models
 {
-    public class CookieConsentViewModel
+    public class CookieConsentVM
     {
         public bool ShowBanner { get; set; }
         public string CookieString { get; set; }
