+++ src/Themes/Clarity/Views/Shared/Analytics.cshtml
@@ -1,4 +1,4 @@
-﻿@model AnalyticsViewModel
+﻿@model AnalyticsVM
 
 @if (!Model.GoogleAnalyticsTrackingID.IsNullOrEmpty())
 {
