+++ src/Themes/Clarity/Views/Shared/CookieConsent.cshtml
@@ -1,4 +1,4 @@
-﻿@model CookieConsentViewModel
+﻿@model CookieConsentVM
 
 @if (Model.ShowBanner)
 {
