+++ src/Themes/Clarity/Views/Shared/Header.cshtml
@@ -1,4 +1,4 @@
-﻿@model HeaderViewModel
+﻿@model HeaderVM
 
 <div class="d-flex flex-column flex-md-row align-items-center px-md-4 pt-2 pb-0 pt-sm-2 pb-sm-0 py-md-2 py-lg-2 py-xl-2 mb-3 bg-white border-bottom box-shadow">
     <h5 class="my-0 mr-md-auto font-weight-normal p-0 px-3 pl-md-0 pl-lg-0">
