+++ src/Themes/Clarity/Views/_ViewImports.cshtml
@@ -2,15 +2,14 @@
 @using Fan.Blog.Helpers
 @using Fan.Blog.Models
 @using Fan.Blog.Models.View
-@using Fan.Membership
 @using Fan.Helpers
+@using Fan.Membership
+@using Fan.Navigation
 @using Fan.Settings
 @using Fan.Themes
 @using Fan.Web
-@using Fan.Web.Models.Auth
 @using Fan.Web.Components
-@using Fan.Web.Models.Home
-@using Fan.Navigation
+@using Fan.Web.Models
 @using Microsoft.AspNetCore.Identity
 @using System
 @using System.Linq
